requirements.txt
Django==3.0.6
djangorestframework==3.10.3
django-allauth
psycopg2-binary==2.8.4